+++ tests/Docs/Models/Subscription.php
@@ -33,8 +33,6 @@ class Subscription extends Model
 
     protected $guarded = [];
 
-    protected $dateFormat = 'Y-m-d H:i:s.u';
-
     /**
      * @var array<string, string>
      */