+++ tests/Fixtures/Models/IndexedPrice.php
@@ -22,8 +22,6 @@ class IndexedPrice extends Model
 
     protected $guarded = [];
 
-    protected $dateFormat = 'Y-m-d H:i:s.u';
-
     /**
      * @return BelongsTo<Product, $this>
      */